Language changes
I more or less rewrote the entire file to use more strict and binding language wich is far more consistent, this is due to the ambiguity of the existing language and its inconsistency. Please revise the format of the entire document to be more in line with the existing style choices but maintain the language used.